requirements.txt
Pillow==7.0.0
numpy==1.18.3
torch==1.9.0
torchvision==0.10.0